add faker, add incident api type with mock
