+++ package.json
@@ -112,6 +112,7 @@
     "@testing-library/react-hooks": "^5.1.0",
     "@testing-library/user-event": "^12.8.3",
     "@types/enzyme": "^3.10.8",
+    "@types/faker": "^5.5.6",
     "@types/invariant": "^2.2.34",
     "@types/jest": "^26.0.21",
     "@types/jsdom": "^16.2.7",
@@ -174,6 +175,7 @@
     "eslint-plugin-testing-library": "^3.10.1",
     "exports-loader": "^2.0.0",
     "express": "^4.17.1",
+    "faker": "^5.5.3",
     "file-loader": "^6.2.0",
     "fork-ts-checker-webpack-plugin": "^6.2.0",
     "html-loader": "^2.1.2",
